[dashboard] migrate embeddable_data_grid test cases to discover functional test (#223452)
`embeddable_data_grid` tests discover session embeddable functionality.
This is owned by discover team and should not be part of dashboard test
suite. This PR consolidates the test cases from embeddable_data_grid to
_saved_search_embeddable.ts.

(cherry picked from commit 05895de3ce2f9cd7b163e496158327ab41ce6867)

# Conflicts:
#	src/platform/test/functional/apps/dashboard/group1/embeddable_data_grid.ts
