+++ src/platform/test/functional/apps/dashboard/group1/index.ts
@@ -37,7 +37,6 @@ export default function ({ getService, loadTestFile }: FtrProviderContext) {
     // a fresh index.
     loadTestFile(require.resolve('./url_field_formatter'));
     loadTestFile(require.resolve('./embeddable_rendering'));
-    loadTestFile(require.resolve('./embeddable_data_grid'));
     loadTestFile(require.resolve('./edit_embeddable_redirects'));
     loadTestFile(require.resolve('./dashboard_unsaved_state'));
     loadTestFile(require.resolve('./dashboard_unsaved_listing'));
